README.md
This is a little side game that I created in my free time using javascript. It's a game where you're presented with two images, a left hand-side and a right-hand side, both of which are filled up with smileys. One side, however, has one extra face than the other. The aim of the game is to click on the extra face.